# libobs: Rework RGB to YUV conversion