### Description
RGB to YUV converison was previously baked into every scale shader, but
this work has been moved to the YUV packing shaders. The scale shaders
now write RGBA instead. In the case where base and output resolutions
are identical, the render texture is forwarded directly to the YUV pack
step, skipping an entire fullscreen pass.

### Motivation and Context
Scale shaders shouldn't really need to know anything about YUV conversions, but the main reason is to reduce unnecessary GPU writes in the common case of unscaled output.

### How Has This Been Tested?

Intel GPA trace inspection:
- [x] D3D/GL - (no-scale, skip draw)
- [x] D3D/GL - Bicubic
- [x] D3D/GL - Bilinear
- [x] D3D/GL - Bilinear Lowres
- [x] D3D/GL - Lanczos
- [x] D3D/GL - 420 output
- [x] D3D/GL - 444 output
- [x] D3D/GL - RGB output
- [x] D3D/GL - NV12 output
- [x] D3D - NV12 output (hardware format)

Intel GPA, SetStablePowerState, Intel HD Graphics 530, NV12

1920x1080, expect win from skipping fullscreen pass

Before:
RGBA -> UYVX: ~321 us
UYVX -> Y: ~480 us
UYVX -> UV: ~127 us

After:
[forward render texture]
RGBA -> Y: ~487 us
RGBA -> UV: ~131 us

1920x1080 -> 1280x720, expect no performance difference

Before:
RGBA -> UYVX: ~268 us
UYVX -> Y: ~209 us
UYVX -> UV: ~57 us

After:
RGBA -> RGBA (rescale): ~268 us
RGBA -> Y: ~210 us
RGBA -> UV: ~58 us

(GPU times aren't ordinarily this stable.)

### Types of changes
- Performance enhancement (non-breaking change which improves efficiency)
- Code cleanup (non-breaking change which makes code smaller or more readable)

### Checklist:
- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.