@jpark37 2019-07-22
I just noticed the CPU conversion path. It doesn't appear to be active, and it's unusable now, so I removed it. Breakpoint-inspected all four output formats I420/I444/NV12/RGB, and they are all behaving as they should.