@kkartaltepe 2019-07-22
It sounds like those two should be two separate commits since it seemed like they were unrelated and your code worked perfectly fine with it in previously. It would make it easier to review both of them as well.